requirements.txt
apache-beam==2.43.0
asttokens==2.0.5
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
crcmod==1.7
debugpy==1.6.2
decorator==5.1.1
dill==0.3.1.1
docopt==0.6.2
entrypoints==0.4
executing==0.9.1
fastavro==1.7.0
fasteners==0.18
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
objsize==0.5.2
orjson==3.8.1
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.12.0
pymongo==3.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==23.2.0
regex==2022.10.31
requests==2.28.1
rsa==4.9
six==1.16.0
stack-data==0.3.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
zstandard==0.19.0
